Removing: Unecessary files.

+++ include/sched_itr.h
@@ -15,11 +15,11 @@
 	 * @name Operations on sched_itr
 	 */
 	/**@{*/
-	extern sched_itr_tt scheditr_create(int, int);
-	extern void         scheditr_destroy(sched_itr_tt);
+	extern sched_itr_tt scheditr_create(unsigned long int, int);
+	extern void scheditr_destroy(sched_itr_tt);
 
-    extern int    scheditr_twork(const_sched_itr_tt);
-    extern int    scheditr_ntasks(const_sched_itr_tt);
+    extern unsigned long int scheditr_twork(const_sched_itr_tt);
+    extern int scheditr_ntasks(const_sched_itr_tt);
 
     /**@}*/
